Heuristic loading of the core shared object. (#18)
When working outside conda, `libxmipp4-core.so` can not be loaded as the
prefix for virtualenv/system python installation is not being added to
the `LD_LIBRARY_PATH`. As we know the installation path for the core
library, we now pre-load the core library with a Python function before
loading the binding (which would fail if `libxmipp4-core.so` cannot be
found in `LD_LIBRARY_PATH`)

Here we search in the following order:
1. `LD_LIBRARY_PATH` (or platform equivalent) succeeds when xmipp4 is
installed with conda or system-wide.
2. In an heuristic installation directory:
  - Linux:  `<python_data_dir>/lib` or `<python_data_dir>/lib64`
  - OSX: `<python_data_dir>/lib`
  - Windows `<python_data_dir>/bin`

---------

Co-authored-by: Martin Salinas <ssalinasmartin@gmail.com>